requirements.txt
odoo_upgrade @ git+https://github.com/odoo/upgrade-util@master
cairosvg==2.7.0
wallee==5.1.0
openupgradelib
